Separated parsing functions into separate internal objects.
